pyproject.toml
[tool.poetry]
name = "pipeline-api"
version = "0.1.0"
description = ""
authors = ["Leandro Amoras <amorasleandro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
tinydb = "^4.8.2"
sqlalchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
